Refactored and updated tags for rune. Added KTCenterFlowLayout library for collection view
